modules/12-computer-vision/03-implementations/python/requirements.txt
numpy
pillow
scikit-learn
pytest
